+++ ChangeLog.md
@@ -2,6 +2,10 @@ ISAAC Changelog
 
 This changelog summarizes changes and fixes which are a part of each revision.  For more details on the fixes, refer tracking numbers where provided, and the git commit history.
 
+* 2017/09/08 - 5.05
+    * Fixes to trim XML import data of Properties, Designations, and supported trimming new Subsets and Types (Property and Relationship) (585147)
+    * Fixes to move SubsetMemberships during MoveFromConceptCode (583082)
+
 * 2017/09/07 - 5.04
     * Fixes for trimming XML input data 
     * Changes to support MoveFromConceptCode via XML import (583082)